requirements-dev.txt
devtools==0.5.1
aiohttp-devtools==0.13.1
watchgod==0.6
Sphinx==2.4.4
sphinx-rtd-theme==0.5.0
m2r==0.2.1